requirements.txt
asgiref                           3.6.0
backports.entry-points-selectable 1.2.0
certifi                           2022.12.7
chardet                           5.1.0
charset-normalizer                3.1.0
distlib                           0.3.6
Django                            3.0.2
django-bootstrap3                 21.2
django-ckeditor                   6.3.2
django-crispy-forms               1.14.0
django-js-asset                   2.0.0
dnspython                         2.3.0
email-validator                   1.3.1
et-xmlfile                        1.1.0
filelock                          3.10.0
idna                              3.4
importlib-metadata                6.1.0
importlib-resources               5.12.0
numpy                             1.24.2
openpyxl                          3.1.2
pandas                            1.5.3
Pillow                            9.4.0
pip                               23.0.1
pipenv                            2023.3.20
platformdirs                      3.1.1
PyMySQL                           1.0.2
python-dateutil                   2.8.2
pytz                              2022.7.1
requests                          2.28.2
setuptools                        67.5.1
six                               1.16.0
some-package                      0.1
sqlparse                          0.4.3
typing_extensions                 4.5.0
urllib3                           1.26.15
virtualenv                        20.21.0
virtualenv-clone                  0.5.7
wheel                             0.38.4
zipp                              3.15.0